Make linter configurable in the package.json itself
Support a `standard-engine` key in the package.json which maps to a
linter package. This will be the selected linter even if other linters
are detected.

Support scoped packages, assuming their `cmd` is the package name
without the scope prefix. Read options based on that unscoped package
name.

For example:

```json
{
  "standard-engine": "@novemberborn/as-i-preach",
  "as-i-preach": {
    "ignore": "path-to-ignore.js"
  }
}
```

Use `resolve-from` to resolve the linter module. This ensures the linter
is indeed an installed package.

Ensure the worker is created with a CWD set to the `projectRoot`. Don't
replace parts of `pathToLinter`, since this won't work reliably with
scoped or globally installed linters.