Update README to better value and explain linter resolution
Rephrase the value proposition of the package. Explain how linters are
discovered and how the new standard-engine setting comes in to that.

Technically linters are discovered if configuration is present, even
if they're not listened in the dependencies. But that's unlikely to work
unless the linter is installed globally and resolvable inside Atom.