Update findOptions() recursion
Simplify stanza, use `name` and `skip` rather than `linterName` and
`skipPackageJson`.

Resolve `package.json` separately from `getOptions()`. This allows
skip handling to be performed when the `package.json` is resolved, which
is a clearer separation of concerns that doing it inside `getOptions()`.

Determine the final `linterName` in `getOptions()`, including supporting
the `name` property in the config stanza.